product page folder rmvd / product card moved to products folder
