+++ frontend/src/components/Product(s)/ProductHeroAd.js
@@ -40,7 +40,7 @@ const ProductHeroAd = ({products}) => {
                 <Col xs={12} className='d-flex justify-content-end global_white-bg'>
                 {/*<Col xs={12} xl={8} className='d-flex justify-content-end global_white-bg'>*/}
                     <Row>
-                        <Col xs={4} className='d-flex align-items-center text-center'>
+                        <Col lg={4} xs={12} className='d-flex align-items-center text-center'>
                             <Row>
                                 <Col xs={12}>
                                     <h1 className='global_blood-red font-weight-bold'>Ad Title</h1>
@@ -56,31 +56,6 @@ const ProductHeroAd = ({products}) => {
 
                     </Row>
                 </Col>
-
-                {/*<Col xs={4} className='text-center d-none d-xl-block'>*/}
-                {/*    <Row>*/}
-                {/*        <Col xs={12} style={indAd} className='d-flex align-items-center mb-1'>*/}
-                {/*            <Row className='d-flex align-items-center'>*/}
-                {/*                <Col xs={6}>*/}
-                {/*                    <h6 className='global_goldenrod'>Lorem ipsum dolor sit amet.</h6>*/}
-                {/*                </Col>*/}
-                {/*                <Col xs={6}>*/}
-                {/*                    <Image src={ad2} className='w-75 rounded' style={indAd}/>*/}
-                {/*                </Col>*/}
-                {/*            </Row>*/}
-                {/*        </Col>*/}
-                {/*        <Col xs={12} style={indAd} className='d-flex align-items-center'>*/}
-                {/*            <Row className='d-flex align-items-center'>*/}
-                {/*                <Col xs={6}>*/}
-                {/*                    <h6 className='global_goldenrod'>Lorem ipsum dolor sit amet.</h6>*/}
-                {/*                </Col>*/}
-                {/*                <Col xs={6}>*/}
-                {/*                    <Image src={ad3} className='w-75 rounded' style={indAd}/>*/}
-                {/*                </Col>*/}
-                {/*            </Row>*/}
-                {/*        </Col>*/}
-                {/*    </Row>*/}
-                {/*</Col>*/}
             </Row>
