+++ frontend/src/components/Product(s)/ProductsImgDisplay.js
@@ -1,7 +1,7 @@
 import React from "react";
 import {Col, Row} from "react-bootstrap";
 import Paginate from "../Paginate";
-import Product from "../ProductPage/Product";
+import Product from "./Product";
 
 const ProductsImgDisplay = ({sideBar, history, match, updateCatProp, updateManufacturerProp, keyword, pages, page, products}) => {
